+++ packages/react/src/use-object.ts
@@ -1,9 +1,14 @@
-import { isAbortError } from '@ai-sdk/provider-utils';
 import {
-  DeepPartial,
   FetchFunction,
+  isAbortError,
+  safeValidateTypes,
+} from '@ai-sdk/provider-utils';
+import {
+  asSchema,
+  DeepPartial,
   isDeepEqualData,
   parsePartialJson,
+  Schema,
 } from '@ai-sdk/ui-utils';
 import { useCallback, useId, useRef, useState } from 'react';
 import useSWR from 'swr';
@@ -21,7 +26,7 @@ export type Experimental_UseObjectOptions<RESULT> = {
   /**
    * A Zod schema that defines the shape of the complete object.
    */
-  schema: z.Schema<RESULT>;
+  schema: z.Schema<RESULT, z.ZodTypeDef, any> | Schema<RESULT>;
 
   /**
    * An unique identifier. If not provided, a random one will be
@@ -169,9 +174,8 @@ function useObject<RESULT, INPUT = any>({
           write(chunk) {
             accumulatedText += chunk;
 
-            const currentObject = parsePartialJson(
-              accumulatedText,
-            ) as DeepPartial<RESULT>;
+            const { value } = parsePartialJson(accumulatedText);
+            const currentObject = value as DeepPartial<RESULT>;
 
             if (!isDeepEqualData(latestObject, currentObject)) {
               latestObject = currentObject;
@@ -185,10 +189,14 @@ function useObject<RESULT, INPUT = any>({
             abortControllerRef.current = null;
 
             if (onFinish != null) {
-              const validationResult = schema.safeParse(latestObject);
+              const validationResult = safeValidateTypes({
+                value: latestObject,
+                schema: asSchema(schema),
+              });
+
               onFinish(
                 validationResult.success
-                  ? { object: validationResult.data, error: undefined }
+                  ? { object: validationResult.value, error: undefined }
                   : { object: undefined, error: validationResult.error },
               );
             }