+++ packages/ui-utils/src/parse-complex-response.ts
@@ -89,6 +89,10 @@ export async function parseComplexResponse({
   for await (const { type, value } of readDataStream(reader, {
     isAborted: () => abortControllerRef?.current === null,
   })) {
+    if (type === 'error') {
+      throw new Error(value);
+    }
+
     if (type === 'text') {
       if (prefixMap['text']) {
         prefixMap['text'] = {
@@ -150,11 +154,13 @@ export async function parseComplexResponse({
 
       partialToolCall.text += value.argsTextDelta;
 
+      const { value: partialArgs } = parsePartialJson(partialToolCall.text);
+
       prefixMap.text!.toolInvocations![partialToolCall.prefixMapIndex] = {
         state: 'partial-call',
         toolCallId: value.toolCallId,
         toolName: partialToolCall.toolName,
-        args: parsePartialJson(partialToolCall.text),
+        args: partialArgs,
       };
 
       // trigger update for streaming by copying adding a update id that changes